requirements.txt
mysql-connector-python 
yagmail 
python-dotenv